Moved static methods to Util class
